feat: Add starting routes layout for clubs and admin
This commit introduces the starting layout proposed in #2. There is no functionality or "business logic" for any of the routes in this commit. These routes include unique pages for individual clubs as well as a collective overview on the admin side. Within these pages, there are functionalities for viewing and modifying specific aspects of each club, including the ability to approve or deny requests.